[wgsl] Combine var/let tests
The var/let validation tests ended up spread over multiple files. Some
were combined var/let tests, some where individual. This CL moves them
all to be `decl,var` and `decl,let` to make it easier to determine what
validation exists.
