+++ src/webgpu/listing_meta.json
@@ -1913,16 +1913,19 @@
   "webgpu:shader,validation,decl,var:binding_point_on_function_var:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:binding_point_on_non_resources:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:binding_point_on_resources:*": { "subcaseMS": 1.000 },
+  "webgpu:shader,validation,decl,var:explicit_access_mode:*": { "subcaseMS": 33.047 },
   "webgpu:shader,validation,decl,var:function_addrspace_at_module_scope:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:function_scope_types:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:handle_initializer:*": { "subcaseMS": 1.000 },
+  "webgpu:shader,validation,decl,var:implicit_access_mode:*": { "subcaseMS": 13.542 },
   "webgpu:shader,validation,decl,var:initializer_kind:*": { "subcaseMS": 1.000 },
+  "webgpu:shader,validation,decl,var:initializer_type:*": { "subcaseMS": 527.537 },
   "webgpu:shader,validation,decl,var:module_scope_initializers:*": { "subcaseMS": 1.000 },
   "webgpu:shader,validation,decl,var:module_scope_types:*": { "subcaseMS": 1.000 },
-  "webgpu:shader,validation,decl,var_access_mode:explicit_access_mode:*": { "subcaseMS": 1.373 },
-  "webgpu:shader,validation,decl,var_access_mode:implicit_access_mode:*": { "subcaseMS": 1.000 },
-  "webgpu:shader,validation,decl,var_access_mode:read_access:*": { "subcaseMS": 1.177 },
-  "webgpu:shader,validation,decl,var_access_mode:write_access:*": { "subcaseMS": 1.154 },
+  "webgpu:shader,validation,decl,var:read_access:*": { "subcaseMS": 20.792 },
+  "webgpu:shader,validation,decl,var:var_access_mode_bad_other_template_contents:*": { "subcaseMS": 54.257 },
+  "webgpu:shader,validation,decl,var:var_access_mode_bad_template_delim:*": { "subcaseMS": 49.467 },
+  "webgpu:shader,validation,decl,var:write_access:*": { "subcaseMS": 21.476 },
   "webgpu:shader,validation,expression,access,array:abstract_array_concrete_index:*": { "subcaseMS": 1.813 },
   "webgpu:shader,validation,expression,access,array:early_eval_errors:*": { "subcaseMS": 312.634 },
   "webgpu:shader,validation,expression,access,array:index_type:*": { "subcaseMS": 505.611 },
@@ -2569,9 +2572,6 @@
   "webgpu:shader,validation,parse,statement_behavior:valid_statements:*": { "subcaseMS": 15.781 },
   "webgpu:shader,validation,parse,switch:parse:*": { "subcaseMS": 16.068 },
   "webgpu:shader,validation,parse,unary_ops:all:*": { "subcaseMS": 1.000 },
-  "webgpu:shader,validation,parse,var_and_let:initializer_type:*": { "subcaseMS": 0.900 },
-  "webgpu:shader,validation,parse,var_and_let:var_access_mode_bad_other_template_contents:*": { "subcaseMS": 4.071 },
-  "webgpu:shader,validation,parse,var_and_let:var_access_mode_bad_template_delim:*": { "subcaseMS": 1.088 },
   "webgpu:shader,validation,parse,while:parse:*": { "subcaseMS": 8.641 },
   "webgpu:shader,validation,shader_io,binding:binding:*": { "subcaseMS": 1.240 },
   "webgpu:shader,validation,shader_io,binding:binding_f16:*": { "subcaseMS": 0.500 },
