+++ src/webgpu/shader/validation/decl/var.spec.ts
@@ -5,6 +5,15 @@ Validation tests for host-shareable types.
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { keysOf } from '../../../../common/util/data_tables.js';
 import { ShaderValidationTest } from '../shader_validation_test.js';
+import { AddressSpace, kAccessModeInfo, kAddressSpaceInfo } from '../../types.js';
+import {
+  explicitSpaceExpander,
+  getVarDeclShader,
+  accessModeExpander,
+  supportsRead,
+  supportsWrite,
+  ShaderStage,
+} from './util.js';
 
 export const g = makeTestGroup(ShaderValidationTest);
 
@@ -573,4 +582,188 @@ g.test('address_space_access_mode')
       ${fdecl}
     }`;
     t.expectCompileResult(shouldPass, code);
+  });
+
+// Address spaces that can hold an i32 variable.
+const kNonHandleAddressSpaces = keysOf(kAddressSpaceInfo).filter(
+  as => as !== 'handle'
+) as AddressSpace[];
+
+g.test('explicit_access_mode')
+  .desc('Validate uses of an explicit access mode on a var declaration')
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(
+    u =>
+      u
+        .combine('addressSpace', kNonHandleAddressSpaces)
+        .combine('explicitSpace', [true, false])
+        // Only keep cases where:
+        //   *if* the address space must be specified on a var decl (e.g. var<private>)
+        //   then the address space will actually be specified in this test case.
+        .filter(t => kAddressSpaceInfo[t.addressSpace].spell !== 'must' || t.explicitSpace)
+        .combine('explicitAccess', [true])
+        .combine('accessMode', keysOf(kAccessModeInfo))
+        .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
+  )
+  .fn(t => {
+    const prog = getVarDeclShader(t.params);
+    const info = kAddressSpaceInfo[t.params.addressSpace];
+
+    const ok =
+      // The address space must be explicitly specified.
+      t.params.explicitSpace &&
+      // The address space must allow an access mode to be spelled, and the
+      // access mode must be in the list of modes for the address space.
+      info.spellAccessMode !== 'never' &&
+      info.accessModes.includes(t.params.accessMode);
+
+    t.expectCompileResult(ok, prog);
+  });
+
+g.test('implicit_access_mode')
+  .desc('Validate an implicit access mode on a var declaration')
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(
+    u =>
+      u
+        .combine('addressSpace', kNonHandleAddressSpaces)
+        .expand('explicitSpace', explicitSpaceExpander)
+        .combine('explicitAccess', [false])
+        .combine('accessMode', [''] as const)
+        .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
+  )
+  .fn(t => {
+    const prog = getVarDeclShader(t.params);
+
+    // 7.3 var Declarations
+    // "The access mode always has a default value,.."
+    const ok = true;
+
+    t.expectCompileResult(ok, prog);
+  });
+
+g.test('read_access')
+  .desc('A variable can be read from when the access mode permits')
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(
+    u =>
+      u
+        .combine('addressSpace', kNonHandleAddressSpaces)
+        .expand('explicitSpace', explicitSpaceExpander)
+        .combine('explicitAccess', [false, true])
+        .expand('accessMode', accessModeExpander)
+        .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
+  )
+  .fn(t => {
+    const prog = getVarDeclShader(t.params, 'let copy = x;');
+    const ok = supportsRead(t.params);
+    t.expectCompileResult(ok, prog);
+  });
+
+g.test('write_access')
+  .desc('A variable can be written to when the access mode permits')
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(
+    u =>
+      u
+        .combine('addressSpace', kNonHandleAddressSpaces)
+        .expand('explicitSpace', explicitSpaceExpander)
+        .combine('explicitAccess', [false, true])
+        .expand('accessMode', accessModeExpander)
+        .combine('stage', ['compute' as ShaderStage]) // Only need to check compute shaders
+  )
+  .fn(t => {
+    const prog = getVarDeclShader(t.params, 'x = 0;');
+    const ok = supportsWrite(t.params);
+    t.expectCompileResult(ok, prog);
+  });
+
+const kTestTypes = [
+  'f32',
+  'i32',
+  'u32',
+  'bool',
+  'vec2<f32>',
+  'vec2<i32>',
+  'vec2<u32>',
+  'vec2<bool>',
+  'vec3<f32>',
+  'vec3<i32>',
+  'vec3<u32>',
+  'vec3<bool>',
+  'vec4<f32>',
+  'vec4<i32>',
+  'vec4<u32>',
+  'vec4<bool>',
+  'mat2x2<f32>',
+  'mat2x3<f32>',
+  'mat2x4<f32>',
+  'mat3x2<f32>',
+  'mat3x3<f32>',
+  'mat3x4<f32>',
+  'mat4x2<f32>',
+  'mat4x3<f32>',
+  'mat4x4<f32>',
+  // [1]: 12 is a random number here. find a solution to replace it.
+  'array<f32, 12>',
+  'array<i32, 12>',
+  'array<u32, 12>',
+  'array<bool, 12>',
+] as const;
+
+g.test('initializer_type')
+  .desc(
+    `
+  If present, the initializer's type must match the store type of the variable.
+  Testing scalars, vectors, and matrices of every dimension and type.
+  TODO: add test for: structs - arrays of vectors and matrices - arrays of different length
+`
+  )
+  .params(u => u.beginSubcases().combine('lhsType', kTestTypes).combine('rhsType', kTestTypes))
+  .fn(t => {
+    const { lhsType, rhsType } = t.params;
+
+    const code = `
+      @fragment
+      fn main() {
+        var a : ${lhsType} = ${rhsType}();
+      }
+    `;
+
+    const expectation = lhsType === rhsType;
+    t.expectCompileResult(expectation, code);
+  });
+
+g.test('var_access_mode_bad_other_template_contents')
+  .desc(
+    'A variable declaration with explicit access mode with varying other template list contents'
+  )
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(u =>
+    u
+      .combine('accessMode', ['read', 'read_write'])
+      .combine('prefix', ['storage,', '', ','])
+      .combine('suffix', [',storage', ',read', ',', ''])
+  )
+  .fn(t => {
+    const prog = `@group(0) @binding(0)
+                  var<${t.params.prefix}${t.params.accessMode}${t.params.suffix}> x: i32;`;
+    const ok = t.params.prefix === 'storage,' && t.params.suffix === '';
+    t.expectCompileResult(ok, prog);
+  });
+
+g.test('var_access_mode_bad_template_delim')
+  .desc('A variable declaration has explicit access mode with varying template list delimiters')
+  .specURL('https://gpuweb.github.io/gpuweb/wgsl/#var-decls')
+  .params(u =>
+    u
+      .combine('accessMode', ['read', 'read_write'])
+      .combine('prefix', ['', '<', '>', ','])
+      .combine('suffix', ['', '<', '>', ','])
+  )
+  .fn(t => {
+    const prog = `@group(0) @binding(0)
+                  var ${t.params.prefix}storage,${t.params.accessMode}${t.params.suffix} x: i32;`;
+    const ok = t.params.prefix === '<' && t.params.suffix === '>';
+    t.expectCompileResult(ok, prog);
   });
